refactor: lib bitmap renamed dune2

+++ packages/pal2bmp/src/command.rs
@@ -8,17 +8,17 @@ pub fn run(config: Cli) -> Result<(), Box<dyn Error>> {
         fs::create_dir_all(parent)?;
     }
 
-    let palette = bitmap::Palette::load(&mut fs::File::open(&config.input_filepath)?)?;
+    let palette = dune2::Palette::load(&mut fs::File::open(&config.input_filepath)?)?;
     let palette_width = 32*16;
     let palette_height = 32*((palette.len() as f32)/16.).ceil() as u32;
 
-    let mut palette_bitmap = bitmap::Bitmap::new(palette_width, palette_height);
+    let mut palette_bitmap = dune2::Bitmap::new(palette_width, palette_height);
 
-    let palette_color_watch_size = bitmap::Size { width: 32, height: 32 };
+    let palette_color_watch_size = dune2::Size { width: 32, height: 32 };
 
     for (i, color) in palette.iter() {
-        let rect = bitmap::Rect::from_point_and_size(
-            32*bitmap::Point {
+        let rect = dune2::Rect::from_point_and_size(
+            32*dune2::Point {
                 x: (i as i32)%16,
                 y: (i as i32)/16,
             },
@@ -26,6 +26,10 @@ pub fn run(config: Cli) -> Result<(), Box<dyn Error>> {
         );
 
         palette_bitmap.fill_rect(&rect, color);
+    }
+
+    if config.output_filepath.exists() && !config.overwrite {
+        return Err("Output file already exists. Use --force to overwrite.".into());
     }
 
     let mut output = fs::File::create(&config.output_filepath)?;
